+++ src/test/java/ai/fullfillment_v2/inbound/InboundStatus.java
@@ -1,0 +1,11 @@
+package ai.fullfillment_v2.inbound;
+
+public enum InboundStatus {
+    ORDER_REQUESTED("발주 요청");
+
+    private final String description;
+
+    InboundStatus(final String description) {
+        this.description = description;
+    }
+}
